requirements.txt
langgraph
langchain-google-genai>=2.0.0
google-auth
langchain-core
langchain-community
ib_insync           # Python wrapper for IBKR
pandas
numpy
ta                  # Technical Analysis library
pydantic
google-cloud-logging
mcp                 # Model Context Protocol SDK
alpaca-py
pandas_ta
fredapi
yfinance
lxml
